1.1.2

+++ core/src/main/java/za/net/hanro50/dischat/core/InfoProvider.java
@@ -5,7 +5,7 @@ public interface InfoProvider {
   public static class Result {
     public int maxPlayers;
     public int onlinePlayerCount;
-    public float tps;
+    public double tps;
     public byte[] icon;
   }
